KAN-646 refactor(domain): extract filter+sort engine into query::filter_sort (#316)
* refactor(domain): extract filter+sort engine into query::filter_sort

operations.rs was 586 lines and conflated three concerns: the
KanbanOperations service contract, the CardListFilter / ArchivedCardListFilter
request shapes, and the in-memory filter+sort engine
(filter_and_sort_cards, count_filtered_cards, passes_filter, etc.).

Move the filter shapes and engine into a new submodule under query/,
which already owns "how to query cards in memory":

    kanban-domain/src/
    ├── operations.rs        # KanbanOperations trait only
    └── query/
        ├── mod.rs           # CardQueryBuilder + re-exports
        ├── filter_sort.rs   # CardListFilter, ArchivedCardListFilter,
        │                    # filter_and_sort_cards, count_filtered_cards
        └── sprint.rs        # unchanged

Consumer mental model after the split:
- service contract -> kanban_domain::KanbanOperations
- request shape    -> kanban_domain::CardListFilter (or query::filter_sort::)
- run a request    -> kanban_domain::filter_and_sort_cards
- count only       -> kanban_domain::count_filtered_cards
- fluent in-memory -> kanban_domain::CardQueryBuilder

Crate-root re-exports preserved, so kanban-service, kanban-tui,
kanban-cli, kanban-mcp need no changes.

operations.rs: 586 -> 442 lines (still over 300 - that's the trait
surface itself; KAN-645 would split the trait by sub-domain).
query/filter_sort.rs: new at 161 lines.

* chore: add changeset

* chore: rewrite changeset as user-facing entry

Original auto-generated body was a diff summary. Replace with a
user-facing note explaining that this is internal cleanup with no
behaviour change, plus the supporting-improvements section.
