+++ crates/kanban-domain/src/operations.rs
@@ -1,155 +1,11 @@
-use crate::search::{CardSearcher, CompositeSearcher};
-use crate::sort::{resolve_sort, sort_cards_in_place};
+use crate::query::filter_sort::{filter_and_sort_cards, ArchivedCardListFilter, CardListFilter};
 use crate::KanbanResult;
 use crate::{
     AmbiguousMatch, ArchivedCard, BatchResolutionCause, BatchResolutionFailure, Board, BoardUpdate,
-    Card, CardStatus, CardSummary, CardUpdate, Column, ColumnUpdate, CreateCardOptions,
-    KanbanError, SortField, SortOrder, Sprint, SprintUpdate,
+    Card, CardSummary, CardUpdate, Column, ColumnUpdate, CreateCardOptions, KanbanError, Sprint,
+    SprintUpdate,
 };
-use std::borrow::Borrow;
-use std::collections::HashSet;
 use uuid::Uuid;
-
-#[derive(Default, Clone)]
-pub struct CardListFilter {
-    pub board_id: Option<Uuid>,
-    pub column_id: Option<Uuid>,
-    /// Any-of sprint membership. Pass `Some([sid].into())` for a single
-    /// sprint, or a multi-element set for the TUI's sprint-chip filter.
-    pub sprint_ids: Option<HashSet<Uuid>>,
-    pub hide_assigned: bool,
-    pub status: Option<CardStatus>,
-    /// `CompositeSearcher::all` semantics; empty string is a no-op.
-    pub search: Option<String>,
-    pub sort: Option<SortField>,
-    pub sort_order: Option<SortOrder>,
-}
-
-#[derive(Default, Clone)]
-pub struct ArchivedCardListFilter {
-    pub board_id: Option<Uuid>,
-    pub sort: Option<SortField>,
-    pub sort_order: Option<SortOrder>,
-}
-
-fn allowed_column_ids(columns: &[Column], board_id: Option<Uuid>) -> Option<HashSet<Uuid>> {
-    board_id.map(|bid| {
-        columns
-            .iter()
-            .filter(|c| c.board_id == bid)
-            .map(|c| c.id)
-            .collect()
-    })
-}
-
-fn build_searcher(filter: &CardListFilter) -> Option<CompositeSearcher> {
-    filter
-        .search
-        .as_deref()
-        .filter(|q| !q.is_empty())
-        .map(|q| CompositeSearcher::all(q.to_string()))
-}
-
-fn passes_filter(
-    card: &Card,
-    allowed_columns: Option<&HashSet<Uuid>>,
-    searcher: Option<&CompositeSearcher>,
-    board: Option<&Board>,
-    sprints: &[Sprint],
-    filter: &CardListFilter,
-) -> bool {
-    if let Some(allowed) = allowed_columns {
-        if !allowed.contains(&card.column_id) {
-            return false;
-        }
-    }
-    if let Some(column_id) = filter.column_id {
-        if card.column_id != column_id {
-            return false;
-        }
-    }
-    if let Some(ref ids) = filter.sprint_ids {
-        if !ids.is_empty() {
-            match card.sprint_id {
-                Some(sid) if ids.contains(&sid) => {}
-                _ => return false,
-            }
-        }
-    }
-    if filter.hide_assigned && card.sprint_id.is_some() {
-        return false;
-    }
-    if let Some(status) = filter.status {
-        if card.status != status {
-            return false;
-        }
-    }
-    if let Some(searcher) = searcher {
-        let Some(board) = board else { return true };
-        if !searcher.matches(card, board, sprints) {
-            return false;
-        }
-    }
-    true
-}
-
-/// Single filter + sort entry point for in-memory card slices, generic
-/// over anything that borrows a `Card` (so archived cards flow through
-/// the same predicate via their `Borrow<Card>` impl).
-pub fn filter_and_sort_cards<T: Borrow<Card> + Clone>(
-    cards: &[T],
-    columns: &[Column],
-    sprints: &[Sprint],
-    board: Option<&Board>,
-    filter: &CardListFilter,
-) -> Vec<T> {
-    let allowed = allowed_column_ids(columns, filter.board_id);
-    let searcher = build_searcher(filter);
-    let mut result: Vec<T> = cards
-        .iter()
-        .filter(|c| {
-            passes_filter(
-                (*c).borrow(),
-                allowed.as_ref(),
-                searcher.as_ref(),
-                board,
-                sprints,
-                filter,
-            )
-        })
-        .cloned()
-        .collect();
-    if let Some((field, order)) = resolve_sort(filter.sort, filter.sort_order, board) {
-        sort_cards_in_place(&mut result, field, order);
-    }
-    result
-}
-
-/// Count-only variant that shares the predicate without allocating a
-/// result vector or sorting. Used by the TUI badge/count render path.
-pub fn count_filtered_cards<T: Borrow<Card>>(
-    cards: &[T],
-    columns: &[Column],
-    sprints: &[Sprint],
-    board: Option<&Board>,
-    filter: &CardListFilter,
-) -> usize {
-    let allowed = allowed_column_ids(columns, filter.board_id);
-    let searcher = build_searcher(filter);
-    cards
-        .iter()
-        .filter(|c| {
-            passes_filter(
-                (*c).borrow(),
-                allowed.as_ref(),
-                searcher.as_ref(),
-                board,
-                sprints,
-                filter,
-            )
-        })
-        .count()
-}
 
 /// Trait ensuring TUI and CLI implement the same operations.
 /// Adding a method here forces both implementations to add it.
